requirement.txt
python-dotenv
pymongo
google-cloud-language